Refactor common 'if a card left your graveyard this turn' condition to a shared class
